Solve four separate memory leaks related to rsdHalInit
Three of the items are local to RsdCpuReferenceImpl and now freed in
destructor after all threads are stopped.
Last one is the RsdHal item itself where the pointer for some reason
was explicitly cleared but not freed. There is no reference counting
but it should be ok to free in Shutdown.

Change-Id: I7832e412d12f4bd7cc728481ae0c782fa57b57e4